requirements.txt
pandas>=1.5
Pillow>=9.5
PyYAML>=6.0
Jinja2>=3.1
folium>=0.14
